views_count
21.04.2022

+++ admin/includes/view_all_posts.php
@@ -1,31 +1,55 @@
 <?php
-
+//POSTS多选项删除或添加
 if(isset($_POST['chekBoxArray'])){
     $chekBoxArrays  = $_POST['chekBoxArray'];
-    foreach($chekBoxArrays as $postValueId ){
-       $bulk_Oprions = $_POST['bulk_Oprions'];
-       switch($bulk_Oprions){
-           case 'published';
+    foreach ($chekBoxArrays as $postValueId) {
+        $bulk_Oprions = $_POST['bulk_Oprions'];
+        switch ($bulk_Oprions) {
+           case 'published':
            $query = "UPDATE posts SET post_status = '$bulk_Oprions' WHERE post_id= {$postValueId}";
-           $update_to_published_status = mysqli_query($conn,$query);
-           confirmQuery($update_to_published_status );
+           $update_to_published_status = mysqli_query($conn, $query);
+           confirmQuery($update_to_published_status);
            break;
-                case 'draft';
+                case 'draft':
                 $query = "UPDATE posts SET post_status = '$bulk_Oprions' WHERE post_id= {$postValueId}";
-                $update_to_draft_status = mysqli_query($conn,$query);
-                confirmQuery($update_to_draft_status );
+                $update_to_draft_status = mysqli_query($conn, $query);
+                confirmQuery($update_to_draft_status);
                 break;
-                    case 'delete';
-                    $query = "UPDATE posts SET post_status = '$bulk_Oprions' WHERE post_id= {$postValueId}";
-                    $update_to_delete_status = mysqli_query($conn,$query);
-                    confirmQuery($update_to_delete_status );
-                    break;      
-       }
+                    case 'delete':
+                    $query = "DELETE FROM posts WHERE post_id= {$postValueId}";
+                    $update_to_delete_status = mysqli_query($conn, $query);
+                    confirmQuery($update_to_delete_status);
+                    break;
+                        case 'clone':
+                        $query = "SELECT * FROM posts WHERE post_id= {$postValueId}";
+                        $select_posts = mysqli_query($conn, $query);
+                        while ($post_row = mysqli_fetch_assoc($select_posts)) {
+                            $post_id = $post_row["post_id"];
+                            $post_author = $post_row["post_author"];
+                            $post_title = $post_row["post_title"];
+                            $post_category_id = $post_row["post_category_id"];
+                            $post_status = $post_row["post_status"];
+                           // $post_content = $post_content["post_content"];
+                            $post_image = $post_row["post_image"];
+                            $post_tag = $post_row["post_tag"];
+                            $post_comment_count = $post_row["post_comment_count"];
+                            $post_date = $post_row["post_date"];
+                        }
+
+                            $query = "INSERT INTO posts(post_category_id,post_title,post_author,
+                            post_image,post_tag,post_date,post_status)";
+                             $query .= "VALUES({$post_category_id},'{$post_title}','{$post_author}','{$post_image}',
+                            '{$post_tag}',now(),'{$post_status}')";
+                            $copy_query = mysqli_query($conn, $query);
+                            if (!$copy_query) {
+                                die("QUERY FAILED".mysqli_error($conn));
+                            }
+                            break; }
     }
 }
 ?>
 
-<form action="" method="POST">
+<form action="" class='text-center' method="POST">
 <table class="=table table-bordered table-hover">
 <div id="bulkOptionContainer" class="col-xs-4">
     <select class="form-control" name="bulk_Oprions" id="">
@@ -33,6 +57,7 @@ if(isset($_POST['chekBoxArray'])){
         <option value="published">Publish</option>
         <option value="draft">Draft</option>
         <option value="delete">Delete</option>
+        <option value="clone">Clone</option>
     </select>
 </div>
 <div class="col-xs-4">
@@ -50,18 +75,19 @@ if(isset($_POST['chekBoxArray'])){
                 <th>Post Status</th>
                 <th>Post Image</th>
                 <th>Post Tags</th>
-                <th>Post Comment Count</th>
+                <th>Comments</th>
                 <th>Post Date</th>
                 <th>View Post</th>
                 <th>Edit</th>
                 <th>Delete</th>
+                <th>Views</th>
             </tr>
         </thead> 
         <tbody>
     <?php
 
-       // global $conn;
-        $postResults = "SELECT * FROM posts";
+       //$conn = getConnection();
+        $postResults = "SELECT * FROM posts ORDER BY post_id DESC";
         $select_posts = mysqli_query($conn,$postResults); 
             
         while ($post_row = mysqli_fetch_assoc($select_posts)) {
@@ -74,7 +100,7 @@ if(isset($_POST['chekBoxArray'])){
             $post_tag = $post_row["post_tag"];
             $post_comment_count = $post_row["post_comment_count"];
             $post_date = $post_row["post_date"];
-            
+            $post_views_count = $post_row["post_views_count"];
             echo "<tr>";
             ?>
 
@@ -93,7 +119,7 @@ if(isset($_POST['chekBoxArray'])){
             while ($categoriesRow = mysqli_fetch_assoc($select_categories)) {
                 $cat_id = $categoriesRow["id"];
                 $cat_title = $categoriesRow["title"];
-                echo "<td>{$cat_title}</td>";
+                echo "<td class='text-center'>{$cat_title}</td>";
             }
             
            
@@ -106,7 +132,8 @@ if(isset($_POST['chekBoxArray'])){
             echo "<td>{$post_date}</td>";
             echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onClick=\" javascript: return confirm('Are you sure you want to delete it?');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a href='posts.php?reset={$post_id}'>{$post_views_count}</a></td>"; 
             echo "</tr>";
             
         }
@@ -125,16 +152,20 @@ if(isset($_GET["delete"])){
     $delete_query = mysqli_query($conn,$query);
     header("Location:posts.php");
 }
+
+
+if (isset($_GET["reset"])) {
+    
+    //  global $conn;
+    $the_post_id = $_GET["reset"];
+    if ($_SERVER['REQUEST'] !=='POST') {
+        $query = "UPDATE posts SET post_views_count = 0 WHERE post_id =".mysqli_real_escape_string($conn, $_GET["reset"])."";
+        $delete_query = mysqli_query($conn, $query);
+        header("Location:posts.php");
+
+
+    }
+}
 ?>
-
-
-
-
-
-
-
-
-
-
